Pass shellcheck
Additionally, make liberal use of ${VAR:?} to catch errors early if VAR is
unset.

+++ src/vm-remove.sh
@@ -4,10 +4,12 @@
 # Remove a VM, deleting its resources permanently
 #
 
-prog_NAME=$(basename $0)
-prog_DIR=$(readlink -f $(dirname $0))
+prog_NAME=$(basename "$0")
+prog_DIR=$(readlink -f "$(dirname "$0")")
 prog_LIBDIR=${prog_DIR}
-. ${prog_LIBDIR}/functions.sh
+# shellcheck source=src/functions.sh
+. "${prog_LIBDIR}/functions.sh"
+# shellcheck source=config.sh.dist
 . /etc/vm/config.sh
 
 prog_SUMMARY="Remove a virtual machine"
@@ -42,19 +44,20 @@ while getopts "Hh" opt; do
             ;;
     esac
 done
-shift $((${OPTIND}-1))
+shift $((OPTIND-1))
 
 [ $# -ne 1 ] && usage
 safe_lookup_vm "$1" || die "No such VM: '$1'"
 vm_is_running "${vm_ID}" && die "Cannot remove running VM: '${vm_ID}'"
-. ${vm_DIR}/config.sh
+# shellcheck disable=SC1090
+. "${vm_DIR}/config.sh"
 
 # XXX --quiet does not silence useless lvremove informational messages,
 # XXX so we redirect it to /dev/null
-lvremove --quiet --yes ${conf_VM_DISK_DEV} >/dev/null \
+lvremove --quiet --yes "${conf_VM_DISK_DEV:?}" >/dev/null \
     || die "Could not remove logical volume"
-rm "${conf_DIR}/by-name/$(cat ${vm_DIR}/name)" || die "rm link failed"
-( cd ${vm_DIR} && rm * ) || die "rm failed"
-rmdir ${vm_DIR} || die "rmdir failed"
+rm "${conf_DIR:?}/by-name/$(cat "${vm_DIR}/name")" || die "rm link failed"
+( cd "${vm_DIR}" && rm -- * ) || die "rm failed"
+rmdir "${vm_DIR}" || die "rmdir failed"
 
 exit 0
